Revert "Revert "feat: 서울시 문화행사 API 연동 및 데이터 동기화 스케줄러 구현 (#223)" (#224)"
This reverts commit 2aed093a9368240c27a85a3d380cdd3aba4dbcd6.
