+++ service-map/src/main/java/com/danburn/map/MapApplication.java
@@ -2,7 +2,9 @@ package com.danburn.map;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
 public class MapApplication {
