+++ service-map/src/main/java/com/danburn/map/dto/response/EventResponse.java
@@ -1,0 +1,4 @@
+package com.danburn.map.dto.response;
+
+public class EventResponse {
+}
